page conf now in backend
