+++ src/AppBundle/Entity/Repository/PageRepository.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace AppBundle\Entity\Repository;
+
+use Doctrine\ORM\EntityRepository;
+
+/**
+ * PageRepository
+ *
+ * This class was generated by the Doctrine ORM. Add your own custom
+ * repository methods below.
+ */
+class PageRepository extends EntityRepository
+{
+    public function findPagesQuery()
+    {
+        return $this->getEntityManager()
+            ->createQuery('SELECT p FROM AppBundle:Page p');
+    }
+
+    public function findActivePagesQuery()
+    {
+        return $this->getEntityManager()
+            ->createQuery('SELECT p FROM AppBundle:Page p WHERE p.is_active = 1');
+    }
+}
