+++ src/AppBundle/Resources/views/Backend/Page/edit.html.twig
@@ -7,7 +7,8 @@
 {% set body_class = "head-black-color" %}
 {% block body_main %}
     <div class="page-header">
-        <h1>{{ 'oktothek.page_create_header'|trans}}</h1>
+        <h1>{{ 'oktothek.page_edit_header'|trans}}</h1>
+        <a href="{{ path('oktothek_backend_page_index') }}" class="btn btn-default">{{ 'oktothek.page_create_back_to_index'|trans }}</a>
     </div>
  {{ form_start(form)}}
     {{ form_row(form.title) }}
