+++ src/AppBundle/Resources/views/Backend/Page/new.html.twig
@@ -8,6 +8,7 @@
 {% block body_main %}
     <div class="page-header">
         <h1>{{ 'oktothek.page_create_header'|trans}}</h1>
+        <a href="{{ path('oktothek_backend_page_index') }}" class="btn btn-default">{{ 'oktothek.page_create_back_to_index'|trans }}</a>
     </div>
  {{ form_start(form)}}
     {{ form_row(form.title) }}
