+++ src/AppBundle/Resources/views/Backend/Slide/index.html.twig
@@ -8,6 +8,7 @@
 <ul class="nav nav-tabs">
   <li role="presentation" class="active"><a href="{{ path('oktothek_backend_slide_index')}}">{{ 'oktothek.backend_slider_header'|trans}}</a></li>
   <li role="presentation"><a href="{{ path('oktothek_backend_news_index')}}">{{ 'oktothek.backend_news_header'|trans }}</a></li>
+  <li role="presentation"><a href="{{ path('oktothek_backend_page_index') }}">{{ 'oktothek.backend_page_header'|trans }}</a></li>
 </ul>
 <div class="page-header">
     <h1>{{ 'oktothek.backend_slider_header'|trans}}</h1>
